Se organiza proyecto para que funcione por pages, rutas, siendo el sidebar parte del layout y estando presente siempre
