+++ components/InterceptionTable/InterceptionTable.tsx
