+++ pages/_app.tsx
@@ -1,6 +1,12 @@
+import { Layout } from '@/layouts'
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return (
+    <Layout>
+        <Component {...pageProps} />
+    </Layout>
+
+  )
 }
