--- pages/data.json
@@ -1,202 +1,0 @@
-[
-    {
-        "Responsabilidad": "Duración evento",
-        "Sala": "x",
-        "Café Teatro": "x",
-        "Plazoleta": "x",
-        "Aula Taller": "x",
-        "Otros": "x"
-    },
-    {
-        "Responsabilidad": "Pulep",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Info para Pulep",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Rider Técnico",
-        "Sala": "x",
-        "Café Teatro": "x",
-        "Plazoleta": "x",
-        "Aula Taller": "x",
-        "Otros": "x"
-    },
-    {
-        "Responsabilidad": "Cronograma",
-        "Sala": "x",
-        "Café Teatro": "x",
-        "Plazoleta": "x",
-        "Aula Taller": "x",
-        "Otros": "x"
-    },
-    {
-        "Responsabilidad": "Plano de Luces",
-        "Sala": "x",
-        "Café Teatro": "x",
-        "Plazoleta": "x",
-        "Aula Taller": "x",
-        "Otros": "x"
-    },
-    {
-        "Responsabilidad": "Boletería",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Repertorio",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Tramite Sayco",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Tramite Acinpro",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Licencia Sayco",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Licencia Acinpro",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Piezas Publicitarias",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Info Transporte",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Info Alimentación",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Ambulancia",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Acomodadores",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Catering",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Tiquetera",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Transporte",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Permiso Gobierno",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Arqued acomodadores",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Liquidacion de boletería",
-        "Sala": "",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Liquidar Eventos (SAYCO, ACINPRO, CAFÉ...)",
-        "Sala": "x",
-        "Café Teatro": "",
-        "Plazoleta": "",
-        "Aula Taller": "",
-        "Otros": ""
-    },
-    {
-        "Responsabilidad": "Encuestas de satisfacción",
-        "Sala": "x",
-        "Café Teatro": "x",
-        "Plazoleta": "x",
-        "Aula Taller": "x",
-        "Otros": "x"
-    }
-]
